// src/types/api.ts
export enum ProofType {
  NUMERIC = "NUMERIC",
  CASK_STRENGTH = "CASK_STRENGTH",
  BARREL_PROOF = "BARREL_PROOF",
  FULL_PROOF = "FULL_PROOF",
}